+++ state/state_test.go
@@ -5,7 +5,6 @@ package state_test
 
 import (
 	"fmt"
-	"path/filepath"
 	"sort"
 	"strconv"
 	"strings"
@@ -21,7 +20,6 @@ import (
 	"github.com/juju/utils"
 	gc "gopkg.in/check.v1"
 	"gopkg.in/juju/charm.v5"
-	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
 	mgotxn "gopkg.in/mgo.v2/txn"
 
@@ -37,7 +35,6 @@ import (
 	"github.com/juju/juju/storage/poolmanager"
 	"github.com/juju/juju/storage/provider"
 	"github.com/juju/juju/storage/provider/registry"
-	"github.com/juju/juju/testcharms"
 	"github.com/juju/juju/testing"
 	"github.com/juju/juju/testing/factory"
 	"github.com/juju/juju/version"
@@ -739,387 +736,6 @@ func (s *StateSuite) TestIsNotFound(c *gc.C) {
 	err2 := errors.NotFoundf("foo")
 	c.Assert(err1, gc.Not(jc.Satisfies), errors.IsNotFound)
 	c.Assert(err2, jc.Satisfies, errors.IsNotFound)
-}
-
-func (s *StateSuite) dummyCharm(c *gc.C, curlOverride string) (ch charm.Charm, curl *charm.URL, storagePath, bundleSHA256 string) {
-	ch = testcharms.Repo.CharmDir("dummy")
-	if curlOverride != "" {
-		curl = charm.MustParseURL(curlOverride)
-	} else {
-		curl = charm.MustParseURL(
-			fmt.Sprintf("local:quantal/%s-%d", ch.Meta().Name, ch.Revision()),
-		)
-	}
-	storagePath = "dummy-1"
-	bundleSHA256 = "dummy-1-sha256"
-	return ch, curl, storagePath, bundleSHA256
-}
-
-func (s *StateSuite) TestAddCharm(c *gc.C) {
-	// Check that adding charms from scratch works correctly.
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "")
-	dummy, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(dummy.URL().String(), gc.Equals, curl.String())
-
-	doc := state.CharmDoc{}
-	err = s.charms.FindId(state.DocID(s.State, curl.String())).One(&doc)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Logf("%#v", doc)
-	c.Assert(doc.URL, gc.DeepEquals, curl)
-}
-
-func (s *StateSuite) TestAddCharmUpdatesPlaceholder(c *gc.C) {
-	// Check that adding charms updates any existing placeholder charm
-	// with the same URL.
-	ch := testcharms.Repo.CharmDir("dummy")
-
-	// Add a placeholder charm.
-	curl := charm.MustParseURL("cs:quantal/dummy-1")
-	err := s.State.AddStoreCharmPlaceholder(curl)
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Add a deployed charm.
-	storagePath := "dummy-1"
-	bundleSHA256 := "dummy-1-sha256"
-	dummy, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(dummy.URL().String(), gc.Equals, curl.String())
-
-	// Charm doc has been updated.
-	var docs []state.CharmDoc
-	err = s.charms.FindId(state.DocID(s.State, curl.String())).All(&docs)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(docs, gc.HasLen, 1)
-	c.Assert(docs[0].URL, gc.DeepEquals, curl)
-	c.Assert(docs[0].StoragePath, gc.DeepEquals, storagePath)
-
-	// No more placeholder charm.
-	_, err = s.State.LatestPlaceholderCharm(curl)
-	c.Assert(err, jc.Satisfies, errors.IsNotFound)
-}
-
-func (s *StateSuite) assertPendingCharmExists(c *gc.C, curl *charm.URL) {
-	// Find charm directly and verify only the charm URL and
-	// PendingUpload are set.
-	doc := state.CharmDoc{}
-	err := s.charms.FindId(state.DocID(s.State, curl.String())).One(&doc)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Logf("%#v", doc)
-	c.Assert(doc.URL, gc.DeepEquals, curl)
-	c.Assert(doc.PendingUpload, jc.IsTrue)
-	c.Assert(doc.Placeholder, jc.IsFalse)
-	c.Assert(doc.Meta, gc.IsNil)
-	c.Assert(doc.Config, gc.IsNil)
-	c.Assert(doc.StoragePath, gc.Equals, "")
-	c.Assert(doc.BundleSha256, gc.Equals, "")
-
-	// Make sure we can't find it with st.Charm().
-	_, err = s.State.Charm(curl)
-	c.Assert(err, jc.Satisfies, errors.IsNotFound)
-}
-
-func (s *StateSuite) TestPrepareLocalCharmUpload(c *gc.C) {
-	// First test the sanity checks.
-	curl, err := s.State.PrepareLocalCharmUpload(charm.MustParseURL("local:quantal/dummy"))
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with revision, got .*")
-	c.Assert(curl, gc.IsNil)
-	curl, err = s.State.PrepareLocalCharmUpload(charm.MustParseURL("cs:quantal/dummy"))
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with local schema, got .*")
-	c.Assert(curl, gc.IsNil)
-
-	// No charm in state, so the call should respect given revision.
-	testCurl := charm.MustParseURL("local:quantal/missing-123")
-	curl, err = s.State.PrepareLocalCharmUpload(testCurl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(curl, gc.DeepEquals, testCurl)
-	s.assertPendingCharmExists(c, curl)
-
-	// Try adding it again with the same revision and ensure it gets bumped.
-	curl, err = s.State.PrepareLocalCharmUpload(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(curl.Revision, gc.Equals, 124)
-
-	// Also ensure the revision cannot decrease.
-	curl, err = s.State.PrepareLocalCharmUpload(curl.WithRevision(42))
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(curl.Revision, gc.Equals, 125)
-
-	// Check the given revision is respected.
-	curl, err = s.State.PrepareLocalCharmUpload(curl.WithRevision(1234))
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(curl.Revision, gc.Equals, 1234)
-}
-
-func (s *StateSuite) TestPrepareStoreCharmUpload(c *gc.C) {
-	// First test the sanity checks.
-	sch, err := s.State.PrepareStoreCharmUpload(charm.MustParseURL("cs:quantal/dummy"))
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with revision, got .*")
-	c.Assert(sch, gc.IsNil)
-	sch, err = s.State.PrepareStoreCharmUpload(charm.MustParseURL("local:quantal/dummy"))
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with cs schema, got .*")
-	c.Assert(sch, gc.IsNil)
-
-	// No charm in state, so the call should respect given revision.
-	testCurl := charm.MustParseURL("cs:quantal/missing-123")
-	sch, err = s.State.PrepareStoreCharmUpload(testCurl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(sch.URL(), gc.DeepEquals, testCurl)
-	c.Assert(sch.IsUploaded(), jc.IsFalse)
-
-	s.assertPendingCharmExists(c, sch.URL())
-
-	// Try adding it again with the same revision and ensure we get the same document.
-	schCopy, err := s.State.PrepareStoreCharmUpload(testCurl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(sch, jc.DeepEquals, schCopy)
-
-	// Now add a charm and try again - we should get the same result
-	// as with AddCharm.
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "cs:precise/dummy-2")
-	sch, err = s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-	schCopy, err = s.State.PrepareStoreCharmUpload(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(sch, jc.DeepEquals, schCopy)
-
-	// Finally, try poking around the state with a placeholder and
-	// bundlesha256 to make sure we do the right thing.
-	curl = curl.WithRevision(999)
-	first := txn.TestHook{
-		Before: func() {
-			err := s.State.AddStoreCharmPlaceholder(curl)
-			c.Assert(err, jc.ErrorIsNil)
-		},
-		After: func() {
-			err := s.charms.RemoveId(state.DocID(s.State, curl.String()))
-			c.Assert(err, jc.ErrorIsNil)
-		},
-	}
-	second := txn.TestHook{
-		Before: func() {
-			err := s.State.AddStoreCharmPlaceholder(curl)
-			c.Assert(err, jc.ErrorIsNil)
-		},
-		After: func() {
-			err := s.charms.UpdateId(state.DocID(s.State, curl.String()), bson.D{{"$set", bson.D{
-				{"bundlesha256", "fake"}},
-			}})
-			c.Assert(err, jc.ErrorIsNil)
-		},
-	}
-	defer state.SetTestHooks(c, s.State, first, second, first).Check()
-
-	_, err = s.State.PrepareStoreCharmUpload(curl)
-	cause := errors.Cause(err)
-	c.Assert(cause, gc.Equals, txn.ErrExcessiveContention)
-}
-
-func (s *StateSuite) TestUpdateUploadedCharm(c *gc.C) {
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "")
-	_, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Test with already uploaded and a missing charms.
-	sch, err := s.State.UpdateUploadedCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, gc.ErrorMatches, fmt.Sprintf("charm %q already uploaded", curl))
-	c.Assert(sch, gc.IsNil)
-	missingCurl := charm.MustParseURL("local:quantal/missing-1")
-	sch, err = s.State.UpdateUploadedCharm(ch, missingCurl, storagePath, "missing")
-	c.Assert(err, jc.Satisfies, errors.IsNotFound)
-	c.Assert(sch, gc.IsNil)
-
-	// Test with with an uploaded local charm.
-	_, err = s.State.PrepareLocalCharmUpload(missingCurl)
-	c.Assert(err, jc.ErrorIsNil)
-	sch, err = s.State.UpdateUploadedCharm(ch, missingCurl, storagePath, "missing")
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(sch.URL(), gc.DeepEquals, missingCurl)
-	c.Assert(sch.Revision(), gc.Equals, missingCurl.Revision)
-	c.Assert(sch.IsUploaded(), jc.IsTrue)
-	c.Assert(sch.IsPlaceholder(), jc.IsFalse)
-	c.Assert(sch.Meta(), gc.DeepEquals, ch.Meta())
-	c.Assert(sch.Config(), gc.DeepEquals, ch.Config())
-	c.Assert(sch.StoragePath(), gc.DeepEquals, storagePath)
-	c.Assert(sch.BundleSha256(), gc.Equals, "missing")
-}
-
-func (s *StateSuite) TestUpdateUploadedCharmEscapesSpecialCharsInConfig(c *gc.C) {
-	// Make sure when we have mongodb special characters like "$" and
-	// "." in the name of any charm config option, we do proper
-	// escaping before storing them and unescaping after loading. See
-	// also http://pad.lv/1308146.
-
-	// Clone the dummy charm and change the config.
-	configWithProblematicKeys := []byte(`
-options:
-  $bad.key: {default: bad, description: bad, type: string}
-  not.ok.key: {description: not ok, type: int}
-  valid-key: {description: all good, type: boolean}
-  still$bad.: {description: not good, type: float}
-  $.$: {description: awful, type: string}
-  ...: {description: oh boy, type: int}
-  just$: {description: no no, type: float}
-`[1:])
-	chDir := testcharms.Repo.ClonedDirPath(c.MkDir(), "dummy")
-	err := utils.AtomicWriteFile(
-		filepath.Join(chDir, "config.yaml"),
-		configWithProblematicKeys,
-		0666,
-	)
-	c.Assert(err, jc.ErrorIsNil)
-	ch, err := charm.ReadCharmDir(chDir)
-	c.Assert(err, jc.ErrorIsNil)
-	missingCurl := charm.MustParseURL("local:quantal/missing-1")
-	storagePath := "dummy-1"
-
-	preparedCurl, err := s.State.PrepareLocalCharmUpload(missingCurl)
-	c.Assert(err, jc.ErrorIsNil)
-	sch, err := s.State.UpdateUploadedCharm(ch, preparedCurl, storagePath, "missing")
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(sch.URL(), gc.DeepEquals, missingCurl)
-	c.Assert(sch.Revision(), gc.Equals, missingCurl.Revision)
-	c.Assert(sch.IsUploaded(), jc.IsTrue)
-	c.Assert(sch.IsPlaceholder(), jc.IsFalse)
-	c.Assert(sch.Meta(), gc.DeepEquals, ch.Meta())
-	c.Assert(sch.Config(), gc.DeepEquals, ch.Config())
-	c.Assert(sch.StoragePath(), gc.DeepEquals, storagePath)
-	c.Assert(sch.BundleSha256(), gc.Equals, "missing")
-}
-
-func (s *StateSuite) assertPlaceholderCharmExists(c *gc.C, curl *charm.URL) {
-	// Find charm directly and verify only the charm URL and
-	// Placeholder are set.
-	doc := state.CharmDoc{}
-	err := s.charms.FindId(state.DocID(s.State, curl.String())).One(&doc)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(doc.URL, gc.DeepEquals, curl)
-	c.Assert(doc.PendingUpload, jc.IsFalse)
-	c.Assert(doc.Placeholder, jc.IsTrue)
-	c.Assert(doc.Meta, gc.IsNil)
-	c.Assert(doc.Config, gc.IsNil)
-	c.Assert(doc.StoragePath, gc.Equals, "")
-	c.Assert(doc.BundleSha256, gc.Equals, "")
-
-	// Make sure we can't find it with st.Charm().
-	_, err = s.State.Charm(curl)
-	c.Assert(err, jc.Satisfies, errors.IsNotFound)
-}
-
-func (s *StateSuite) TestLatestPlaceholderCharm(c *gc.C) {
-	// Add a deployed charm
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "cs:quantal/dummy-1")
-	_, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Deployed charm not found.
-	_, err = s.State.LatestPlaceholderCharm(curl)
-	c.Assert(err, jc.Satisfies, errors.IsNotFound)
-
-	// Add a charm reference
-	curl2 := charm.MustParseURL("cs:quantal/dummy-2")
-	err = s.State.AddStoreCharmPlaceholder(curl2)
-	c.Assert(err, jc.ErrorIsNil)
-	s.assertPlaceholderCharmExists(c, curl2)
-
-	// Use a URL with an arbitrary rev to search.
-	curl = charm.MustParseURL("cs:quantal/dummy-23")
-	pending, err := s.State.LatestPlaceholderCharm(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(pending.URL(), gc.DeepEquals, curl2)
-	c.Assert(pending.IsPlaceholder(), jc.IsTrue)
-	c.Assert(pending.Meta(), gc.IsNil)
-	c.Assert(pending.Config(), gc.IsNil)
-	c.Assert(pending.StoragePath(), gc.Equals, "")
-	c.Assert(pending.BundleSha256(), gc.Equals, "")
-}
-
-func (s *StateSuite) TestAddStoreCharmPlaceholderErrors(c *gc.C) {
-	ch := testcharms.Repo.CharmDir("dummy")
-	curl := charm.MustParseURL(
-		fmt.Sprintf("local:quantal/%s-%d", ch.Meta().Name, ch.Revision()),
-	)
-	err := s.State.AddStoreCharmPlaceholder(curl)
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with cs schema, got .*")
-
-	curl = charm.MustParseURL("cs:quantal/dummy")
-	err = s.State.AddStoreCharmPlaceholder(curl)
-	c.Assert(err, gc.ErrorMatches, "expected charm URL with revision, got .*")
-}
-
-func (s *StateSuite) TestAddStoreCharmPlaceholder(c *gc.C) {
-	curl := charm.MustParseURL("cs:quantal/dummy-1")
-	err := s.State.AddStoreCharmPlaceholder(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	s.assertPlaceholderCharmExists(c, curl)
-
-	// Add the same one again, should be a no-op
-	err = s.State.AddStoreCharmPlaceholder(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	s.assertPlaceholderCharmExists(c, curl)
-}
-
-func (s *StateSuite) assertAddStoreCharmPlaceholder(c *gc.C) (*charm.URL, *charm.URL, *state.Charm) {
-	// Add a deployed charm
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "cs:quantal/dummy-1")
-	dummy, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Add a charm placeholder
-	curl2 := charm.MustParseURL("cs:quantal/dummy-2")
-	err = s.State.AddStoreCharmPlaceholder(curl2)
-	c.Assert(err, jc.ErrorIsNil)
-	s.assertPlaceholderCharmExists(c, curl2)
-
-	// Deployed charm is still there.
-	existing, err := s.State.Charm(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(existing, jc.DeepEquals, dummy)
-
-	return curl, curl2, dummy
-}
-
-func (s *StateSuite) TestAddStoreCharmPlaceholderLeavesDeployedCharmsAlone(c *gc.C) {
-	s.assertAddStoreCharmPlaceholder(c)
-}
-
-func (s *StateSuite) TestAddStoreCharmPlaceholderDeletesOlder(c *gc.C) {
-	curl, curlOldRef, dummy := s.assertAddStoreCharmPlaceholder(c)
-
-	// Add a new charm placeholder
-	curl3 := charm.MustParseURL("cs:quantal/dummy-3")
-	err := s.State.AddStoreCharmPlaceholder(curl3)
-	c.Assert(err, jc.ErrorIsNil)
-	s.assertPlaceholderCharmExists(c, curl3)
-
-	// Deployed charm is still there.
-	existing, err := s.State.Charm(curl)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(existing, jc.DeepEquals, dummy)
-
-	// Older charm placeholder is gone.
-	doc := state.CharmDoc{}
-	err = s.charms.FindId(curlOldRef).One(&doc)
-	c.Assert(err, gc.Equals, mgo.ErrNotFound)
-}
-
-func (s *StateSuite) TestAllCharms(c *gc.C) {
-	// Add a deployed charm
-	ch, curl, storagePath, bundleSHA256 := s.dummyCharm(c, "cs:quantal/dummy-1")
-	sch, err := s.State.AddCharm(ch, curl, storagePath, bundleSHA256)
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Add a charm reference
-	curl2 := charm.MustParseURL("cs:quantal/dummy-2")
-	err = s.State.AddStoreCharmPlaceholder(curl2)
-	c.Assert(err, jc.ErrorIsNil)
-
-	charms, err := s.State.AllCharms()
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(charms, gc.HasLen, 2)
-
-	c.Assert(charms[0], gc.DeepEquals, sch)
-	c.Assert(charms[1].URL(), gc.DeepEquals, curl2)
 }
 
 func (s *StateSuite) AssertMachineCount(c *gc.C, expect int) {
